JSON
Might be tired from a long day of knowledge, but I had a harder time
than I feel I should having to come up with the JSON syntax based on the
description. I think I am at least close, but that really seemed more
difficult than I feel it aught to have. Shrugs...we'll see how I feel
when this comes back around the bend.

Fist of five:
maybe a 2 for comfort
and a 3 for clarity - i get the concept but not sure I could implement
my way out of a paper bag